api/pyproject.toml
[project]
name = "fastapistart"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.13.1",
    "argon2-cffi>=25.1.0",
    "asyncpg>=0.30.0",
    "black>=24.0.0",
    "celery[redis]>=5.3.0",
    "fastapi>=0.110.0",
    "flake8>=7.0.0",
    "flake8-bugbear>=24.12.12",
    "flake8-docstrings>=1.7.0",
    "httpx>=0.26.0",
    "isort>=5.13.0",
    "mypy>=1.8.0",
    "passlib[bcrypt]>=1.7.4",
    "pre-commit>=3.6.0",
    "psycopg2-binary>=2.9.9",
    "pydantic-settings>=2.1.0",
    "pydantic[email]>=2.6.0",
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.6",
    "redis>=5.0.0",
    "requests>=2.32.5",
    "ruff>=0.2.0",
    "sqlalchemy>=2.0.25",
    "uvicorn[standard]>=0.27.0",
]
